+++ lib/parse/query/constraint.rb
@@ -120,6 +120,12 @@ module Parse
       d = d.pointer if d.respond_to?(:pointer) #simplified query object
       #d = d.pointer if d.is_a?(Parse::Object) #simplified query object
       d = d.source if d.is_a?(Regexp)
+      #  d = d.compile
+      if d.is_a?(Parse::Query)
+        compiled = d.compile(false)
+        compiled["className"] = d.table
+        d = compiled
+      end
       d
     end